+++ cmd/migrate_command_test.go
@@ -13,60 +13,53 @@ import (
 	"github.com/Azure/azapi2azurerm/azurerm"
 	"github.com/Azure/azapi2azurerm/cmd"
 	"github.com/Azure/azapi2azurerm/tf"
+	"github.com/Azure/azapi2azurerm/types"
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclwrite"
 	"github.com/mitchellh/cli"
 )
 
 func TestMigrate_basic(t *testing.T) {
-	migrateTestCase(t, basic())
+	migrateTestCase(t, basic(), "azurerm")
 }
 
 func TestMigrate_foreach(t *testing.T) {
-	migrateTestCase(t, foreach())
+	migrateTestCase(t, foreach(), "azurerm")
 }
 
 func TestMigrate_nestedBlock(t *testing.T) {
-	migrateTestCase(t, nestedBlock())
+	migrateTestCase(t, nestedBlock(), "azurerm")
 }
 
 func TestMigrate_count(t *testing.T) {
-	migrateTestCase(t, count())
+	migrateTestCase(t, count(), "azurerm")
 }
 
 func TestMigrate_nestedBlockUpdate(t *testing.T) {
-	migrateTestCase(t, nestedBlockUpdate())
+	migrateTestCase(t, nestedBlockUpdate(), "azurerm")
 }
 
 func TestMigrate_metaArguments(t *testing.T) {
-	migrateTestCase(t, metaArguments())
+	migrateTestCase(t, metaArguments(), "azurerm")
 }
 
-func TestMigrate_basic_payload(t *testing.T) {
-	migrateTestCase(t, basic_payload())
+func TestMigrate_basicAzureRM(t *testing.T) {
+	migrateTestCase(t, basicAzureRM(), "azapi")
 }
 
-func TestMigrate_foreach_payload(t *testing.T) {
-	migrateTestCase(t, foreach_payload())
+func TestMigrate_foreachAzureRM(t *testing.T) {
+	migrateTestCase(t, foreachAzureRM(), "azapi")
 }
 
-func TestMigrate_nestedBlock_payload(t *testing.T) {
-	migrateTestCase(t, nestedBlock_payload())
+func TestMigrate_countAzureRM(t *testing.T) {
+	migrateTestCase(t, countAzureRM(), "azapi")
 }
 
-func TestMigrate_count_payload(t *testing.T) {
-	migrateTestCase(t, count_payload())
+func TestMigrate_metaArgumentsAzureRM(t *testing.T) {
+	migrateTestCase(t, metaArgumentsAzureRM(), "azapi")
 }
 
-func TestMigrate_nestedBlockUpdate_payload(t *testing.T) {
-	migrateTestCase(t, nestedBlockUpdate_payload())
-}
-
-func TestMigrate_metaArguments_payload(t *testing.T) {
-	migrateTestCase(t, metaArguments_payload())
-}
-
-func migrateTestCase(t *testing.T, content string, ignore ...string) {
+func migrateTestCase(t *testing.T, content string, targetProvider string, ignore ...string) {
 	if len(os.Getenv("TF_ACC")) == 0 {
 		t.Skipf("Set `TF_ACC=true` to enable this test")
 	}
@@ -87,6 +80,7 @@ provider "azurerm" {
 	if err != nil {
 		t.Fatal(err)
 	}
+	t.Logf("dir: %s", dir)
 	terraform, err := tf.NewTerraform(dir, false)
 	if err != nil {
 		t.Fatal(err)
@@ -123,27 +117,37 @@ provider "azurerm" {
 	if err != nil {
 		log.Fatal(err)
 	}
-	resources := terraform.ListGenericResources(p)
-	updateResources := terraform.ListGenericUpdateResources(p)
-	for i, r := range resources {
-		resourceId := r.Instances[0].ResourceId
-		resourceTypes, _, err := azurerm.GetAzureRMResourceType(resourceId)
-		if err != nil {
-			t.Fatal(err)
+	allResources := types.ListResourcesFromPlan(p)
+	for index, r := range allResources {
+		switch v := r.(type) {
+		case *types.AzapiResource:
+			resourceId := v.Instances[0].ResourceId
+			resourceTypes, _, err := azurerm.GetAzureRMResourceType(resourceId)
+			if err != nil {
+				t.Fatal(err)
+			}
+			v.ResourceType = resourceTypes[0]
+			allResources[index] = r
+		case *types.AzapiUpdateResource:
+			resourceId := v.Id
+			resourceTypes, _, err := azurerm.GetAzureRMResourceType(resourceId)
+			if err != nil {
+				t.Fatal(err)
+			}
+			v.ResourceType = resourceTypes[0]
+			allResources[index] = r
 		}
-		resources[i].ResourceType = resourceTypes[0]
 	}
-	for i, r := range updateResources {
-		resourceId := r.Id
-		resourceTypes, _, err := azurerm.GetAzureRMResourceType(resourceId)
-		if err != nil {
-			t.Fatal(err)
+
+	migrateResources := make([]types.AzureResource, 0)
+	for _, r := range allResources {
+		if r.TargetProvider() == targetProvider {
+			migrateResources = append(migrateResources, r)
 		}
-		updateResources[i].ResourceType = resourceTypes[0]
 	}
-	migrateCommand := cmd.MigrateCommand{Ui: ui}
-	migrateCommand.MigrateGenericResource(terraform, resources)
-	migrateCommand.MigrateGenericUpdateResource(terraform, updateResources)
+
+	migrateCommand := cmd.MigrateCommand{Ui: ui, TargetProvider: targetProvider}
+	migrateCommand.MigrateResources(terraform, migrateResources)
 
 	// check generic resources are migrated
 	config, err := os.ReadFile(filename)
@@ -159,11 +163,8 @@ provider "azurerm" {
 		t.Fatal("expect a valid file, but got nil")
 	}
 	migratedSet := make(map[string]bool)
-	for _, r := range resources {
+	for _, r := range migrateResources {
 		migratedSet[r.OldAddress(nil)] = true
-	}
-	for _, r := range updateResources {
-		migratedSet[r.OldAddress()] = true
 	}
 	ignoreSet := make(map[string]bool)
 	for _, r := range ignore {
@@ -187,7 +188,7 @@ provider "azurerm" {
 	if err != nil {
 		t.Fatal(err)
 	}
-	if changes := tf.GetChanges(plan); len(changes) != 0 {
+	if changes := GetChanges(plan); len(changes) != 0 {
 		t.Fatalf("expect no plan-diff, but got %v", changes)
 	}
 
@@ -213,7 +214,7 @@ func tempDir(t *testing.T) string {
 }
 
 func randomResourceName() string {
-	return fmt.Sprintf("acctest%d", rand.Intn(10000))
+	return fmt.Sprintf("acctest%d", rand.Intn(100000))
 }
 
 func template() string {
@@ -333,7 +334,6 @@ func foreach() string {
 data "azurerm_client_config" "current" {
 }
 
-
 variable "accounts" {
   type = map(any)
   default = {
@@ -377,11 +377,12 @@ func nestedBlock() string {
 %s
 
 resource "azurerm_storage_account" "test" {
-  name                     = "%s"
-  resource_group_name      = azurerm_resource_group.test.name
-  location                 = azurerm_resource_group.test.location
-  account_tier             = "Standard"
-  account_replication_type = "GRS"
+  name                            = "%s"
+  resource_group_name             = azurerm_resource_group.test.name
+  location                        = azurerm_resource_group.test.location
+  account_tier                    = "Standard"
+  account_replication_type        = "GRS"
+  allow_nested_items_to_be_public = false
 }
 
 variable "description" {
@@ -565,7 +566,7 @@ resource "azapi_resource" "test1" {
 `, template(), randomResourceName(), randomResourceName())
 }
 
-func basic_payload() string {
+func basicAzureRM() string {
 	return fmt.Sprintf(`
 %s
 data "azurerm_client_config" "current" {
@@ -585,76 +586,37 @@ locals {
   AutomationSku = "Basic"
 }
 
-resource "azapi_resource" "test" {
-  name                   = var.AutomationName
-  parent_id              = azurerm_resource_group.test.id
-  type                   = "Microsoft.Automation/automationAccounts@2020-01-13-preview"
-  response_export_values = ["name", "identity", "properties.sku"]
-
-  location = azurerm_resource_group.test.location
+resource "azurerm_automation_account" "automationAccount" {
+  name                = var.AutomationName
+  resource_group_name = azurerm_resource_group.test.name
+  location            = azurerm_resource_group.test.location
+  sku_name            = "Basic"
   identity {
     type = "SystemAssigned"
   }
-
-  body = {
-    properties = {
-      sku = {
-        name = local.AutomationSku
-      }
-    }
-  }
 }
 
-resource "azapi_resource" "test2" {
-  name      = "${var.AutomationName}another"
-  parent_id = azurerm_resource_group.test.id
-  type      = "Microsoft.Automation/automationAccounts@2020-01-13-preview"
-  location  = azurerm_resource_group.test.location
-  body = {
-    properties = {
-      sku = {
-        name = azapi_resource.test.output.properties.sku.name
-      }
-    }
-  }
-}
-
-resource "azurerm_automation_account" "test1" {
-  location            = "westeurope"
-  name                = "%s"
+resource "azurerm_automation_account" "automationAccount2" {
+  name                = "${var.AutomationName}another"
   resource_group_name = azurerm_resource_group.test.name
-  sku_name            = "Basic"
-}
-
-resource "azapi_update_resource" "test" {
-  resource_id            = azurerm_automation_account.test1.id
-  type                   = "Microsoft.Automation/automationAccounts@2020-01-13-preview"
-  response_export_values = ["properties.sku"]
-  body = {
-    tags = {
-      key = var.Label
-    }
+  location            = azurerm_resource_group.test.location
+  sku_name            = azurerm_automation_account.automationAccount.sku_name
+  identity {
+    type = "SystemAssigned"
   }
 }
 
 output "accountName" {
-  value = azapi_resource.test.output.name
+  value = azurerm_automation_account.automationAccount.name
 }
 
-output "patchAccountSKU" {
-  value = azapi_update_resource.test.output.properties.sku.name
-}
-`, template(), randomResourceName(), randomResourceName())
+
+`, template(), randomResourceName())
 }
 
-func foreach_payload() string {
+func foreachAzureRM() string {
 	return fmt.Sprintf(`
 %s
-
-data "azurerm_client_config" "current" {
-}
-
-
 variable "accounts" {
   type = map(any)
   default = {
@@ -669,23 +631,13 @@ variable "accounts" {
   }
 }
 
-
-resource "azapi_resource" "test" {
-  name      = "henglu${each.value.name}"
-  parent_id = azurerm_resource_group.test.id
-  type      = "Microsoft.Automation/automationAccounts@2020-01-13-preview"
-
-  location = azurerm_resource_group.test.location
+resource "azurerm_automation_account" "test" {
+  name                = "henglu${each.value.name}"
+  resource_group_name = azurerm_resource_group.test.name
+  location            = azurerm_resource_group.test.location
+  sku_name            = each.value.sku
   identity {
     type = "SystemAssigned"
-  }
-
-  body = {
-    properties = {
-      sku = {
-        name = each.value.sku
-      }
-    }
   }
 
   for_each = var.accounts
@@ -693,152 +645,37 @@ resource "azapi_resource" "test" {
 `, template(), randomResourceName(), randomResourceName())
 }
 
-func nestedBlock_payload() string {
+func countAzureRM() string {
 	return fmt.Sprintf(`
 %s
 
-resource "azurerm_storage_account" "test" {
-  name                     = "%s"
-  resource_group_name      = azurerm_resource_group.test.name
-  location                 = azurerm_resource_group.test.location
-  account_tier             = "Standard"
-  account_replication_type = "GRS"
-}
-
-variable "description" {
-  type    = string
-  default = "this is my desc"
-}
-
-variable "defName" {
-  type    = string
-  default = "def1"
-}
-
-resource "azapi_resource" "test" {
-  name      = "%s"
-  parent_id = azurerm_resource_group.test.id
-  type      = "Microsoft.Network/serviceEndpointPolicies@2020-11-01"
-
-  body = {
-    location = "westeurope"
-    tags     = {}
-    properties = {
-      serviceEndpointPolicyDefinitions = [
-        {
-          name = var.defName
-          properties = {
-            service     = "Microsoft.Storage"
-            description = var.description
-            serviceResources = [
-              azurerm_storage_account.test.id,
-              azurerm_resource_group.test.id
-            ]
-          }
-        }
-      ]
-    }
-  }
-}
-`, template(), randomResourceName(), randomResourceName())
-}
-
-func count_payload() string {
-	return fmt.Sprintf(`
-%s
-
-resource "azapi_resource" "test" {
-  name      = "%s${count.index}"
-  parent_id = azurerm_resource_group.test.id
-  type      = "Microsoft.Automation/automationAccounts@2020-01-13-preview"
-  location  = azurerm_resource_group.test.location
-  body = {
-    properties = {
-      sku = {
-        name = "Basic"
-      }
-    }
-  }
-
-  count = 2
-}
-`, template(), randomResourceName())
-}
-
-func nestedBlockUpdate_payload() string {
-	return fmt.Sprintf(`
-%s
-
-resource "azurerm_container_registry" "test" {
-  name                = "%s"
+resource "azurerm_automation_account" "test" {
+  name                = "%s${count.index}"
   resource_group_name = azurerm_resource_group.test.name
   location            = azurerm_resource_group.test.location
-  sku                 = "Premium"
-  admin_enabled       = false
-
-  network_rule_set = [{
-    default_action = "Deny"
-    ip_rule = [{
-      action   = "Allow"
-      ip_range = "1.1.1.1/32"
-      }, {
-      action   = "Allow"
-      ip_range = "8.8.8.8/32"
-    }]
-    virtual_network = []
-  }]
-}
-
-variable "action" {
-  type    = string
-  default = "Allow"
-}
-
-resource "azapi_update_resource" "test" {
-  resource_id = azurerm_container_registry.test.id
-  type        = "Microsoft.ContainerRegistry/registries@2019-05-01"
-  body = {
-    properties = {
-      networkRuleSet = {
-        defaultAction = "Deny"
-        ipRules = [
-          {
-            action = var.action
-            value  = "7.7.7.7"
-          },
-          {
-            action = var.action
-            value  = "2.2.2.2"
-          }
-        ]
-      }
-    }
-  }
-}
-`, template(), randomResourceName())
-}
-
-func metaArguments_payload() string {
-	return fmt.Sprintf(`
-%s
-
-resource "azapi_resource" "test" {
-  name                   = "%s"
-  parent_id              = azurerm_resource_group.test.id
-  type                   = "Microsoft.Automation/automationAccounts@2020-01-13-preview"
-  response_export_values = ["name", "identity", "properties.sku"]
-
-  location = azurerm_resource_group.test.location
+  sku_name            = "Basic"
   identity {
     type = "SystemAssigned"
   }
 
-  body = {
-    properties = {
-      sku = {
-        name = "Basic"
-      }
-    }
+  count = 2
+}
+
+
+`, template(), randomResourceName())
+}
+
+func metaArgumentsAzureRM() string {
+	return fmt.Sprintf(`
+%s
+
+resource "azurerm_automation_account" "automationAccount" {
+  name                = "acctest%s"
+  resource_group_name = azurerm_resource_group.test.name
+  location            = azurerm_resource_group.test.location
+  sku_name            = "Basic"
+  identity {
+    type = "SystemAssigned"
   }
 
   depends_on = [azurerm_resource_group.test]
@@ -852,36 +689,5 @@ resource "azapi_resource" "test" {
     command = "echo the resource id is ${self.id}"
   }
 }
-
-
-resource "azapi_resource" "test1" {
-  name      = "%s"
-  parent_id = azurerm_resource_group.test.id
-  type      = "Microsoft.Automation/automationAccounts@2020-01-13-preview"
-
-  location = azurerm_resource_group.test.location
-  identity {
-    type = "SystemAssigned"
-  }
-
-  body = {
-    properties = {
-      sku = {
-        name = "Basic"
-      }
-    }
-  }
-
-  depends_on = [azurerm_resource_group.test, azapi_resource.test]
-
-  lifecycle {
-    create_before_destroy = false
-    prevent_destroy       = false
-  }
-
-  provisioner "local-exec" {
-    command = "echo the resource id is ${self.id}"
-  }
-}
-`, template(), randomResourceName(), randomResourceName())
+`, template(), randomResourceName())
 }